+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 - Replace unsupported IE javascript functions with underscore equivalent ones
 - Backoffice performance optimisations
+- Ensure properties have unique alias and id
 
 ## 0.2.0 - 2016-07-04
 ### Features